Examples and Readme
