+++ examples/scantest.java
