--- registerdemo.java
@@ -1,15 +1,0 @@
-void main() {
-    println("""
-            Clear: ~
-            Markdown: # Register Test
-            Register: Reverse java --enable-preview external.java
-            Reverse: Hello World
-            | Markdown
-            Register[skipId]: Wc wc
-            Wc:
-            Hello World
-            Test
-            ~~~
-            | Html
-            """);
-}
